clean up robot driver
