--- examples/test_robotdriver/RobotDriver.kicad_prl
@@ -1,77 +1,0 @@
-{
-  "board": {
-    "active_layer": 37,
-    "active_layer_preset": "",
-    "auto_track_width": true,
-    "hidden_nets": [],
-    "high_contrast_mode": 0,
-    "net_color_mode": 1,
-    "opacity": {
-      "pads": 1.0,
-      "tracks": 1.0,
-      "vias": 1.0,
-      "zones": 0.6
-    },
-    "ratsnest_display_mode": 0,
-    "selection_filter": {
-      "dimensions": true,
-      "footprints": true,
-      "graphics": true,
-      "keepouts": true,
-      "lockedItems": true,
-      "otherItems": true,
-      "pads": true,
-      "text": true,
-      "tracks": true,
-      "vias": true,
-      "zones": true
-    },
-    "visible_items": [
-      0,
-      1,
-      2,
-      3,
-      4,
-      5,
-      6,
-      7,
-      8,
-      9,
-      10,
-      11,
-      12,
-      13,
-      14,
-      15,
-      16,
-      17,
-      18,
-      19,
-      20,
-      21,
-      22,
-      23,
-      24,
-      25,
-      26,
-      27,
-      28,
-      29,
-      30,
-      32,
-      33,
-      34,
-      35,
-      36
-    ],
-    "visible_layers": "003ffff_80000001",
-    "zone_display_mode": 1
-  },
-  "meta": {
-    "filename": "RobotDriver.kicad_prl",
-    "version": 3
-  },
-  "project": {
-    "files": []
-  }
-}
