--- examples/test_robotdriver/fp-lib-table
@@ -1,3 +1,0 @@
-(fp_lib_table
-  (lib (name edg)(type KiCad)(uri ${KIPRJMOD}/../edg.pretty)(options "")(descr ""))
-)
