--- examples/test_robotdriver/gerbers/RobotDriver-B_Mask.gbr
@@ -1,51 +1,0 @@
-G04 #@! TF.GenerationSoftware,KiCad,Pcbnew,(6.0.6-0)*
-G04 #@! TF.CreationDate,2022-08-30T17:18:48-07:00*
-G04 #@! TF.ProjectId,RobotDriver,526f626f-7444-4726-9976-65722e6b6963,rev?*
-G04 #@! TF.SameCoordinates,Original*
-G04 #@! TF.FileFunction,Soldermask,Bot*
-G04 #@! TF.FilePolarity,Negative*
-%FSLAX46Y46*%
-G04 Gerber Fmt 4.6, Leading zero omitted, Abs format (unit mm)*
-G04 Created by KiCad (PCBNEW (6.0.6-0)) date 2022-08-30 17:18:48*
-%MOMM*%
-%LPD*%
-G01*
-G04 APERTURE LIST*
-%ADD10C,0.650000*%
-%ADD11C,1.152000*%
-%ADD12R,1.700000X1.700000*%
-%ADD13O,1.700000X1.700000*%
-G04 APERTURE END LIST*
-D10*
-X100325000Y-52500000D02*
-G75*
-G03*
-X100325000Y-52500000I-325000J0D01*
-G01*
-X213825000Y-85000000D02*
-G75*
-G03*
-X213825000Y-85000000I-325000J0D01*
-G01*
-X100325000Y-117500000D02*
-G75*
-G03*
-X100325000Y-117500000I-325000J0D01*
-G01*
-D11*
-X100000000Y-52500000D03*
-X213500000Y-85000000D03*
-D12*
-X101250000Y-57450000D03*
-D13*
-X101250000Y-59990000D03*
-X101250000Y-62530000D03*
-X101250000Y-65070000D03*
-D11*
-X100000000Y-117500000D03*
-D12*
-X121000000Y-59540000D03*
-D13*
-X121000000Y-57000000D03*
-X121000000Y-54460000D03*
-M02*
