--- examples/test_robotdriver/gerbers/RobotDriver-B_Silkscreen.gbr
@@ -1,15 +1,0 @@
-G04 #@! TF.GenerationSoftware,KiCad,Pcbnew,(6.0.6-0)*
-G04 #@! TF.CreationDate,2022-08-30T17:18:48-07:00*
-G04 #@! TF.ProjectId,RobotDriver,526f626f-7444-4726-9976-65722e6b6963,rev?*
-G04 #@! TF.SameCoordinates,Original*
-G04 #@! TF.FileFunction,Legend,Bot*
-G04 #@! TF.FilePolarity,Positive*
-%FSLAX46Y46*%
-G04 Gerber Fmt 4.6, Leading zero omitted, Abs format (unit mm)*
-G04 Created by KiCad (PCBNEW (6.0.6-0)) date 2022-08-30 17:18:48*
-%MOMM*%
-%LPD*%
-G01*
-G04 APERTURE LIST*
-G04 APERTURE END LIST*
-M02*
