--- examples/test_robotdriver/gerbers/RobotDriver-Edge_Cuts.gbr
@@ -1,28 +1,0 @@
-G04 #@! TF.GenerationSoftware,KiCad,Pcbnew,(6.0.6-0)*
-G04 #@! TF.CreationDate,2022-08-30T17:18:49-07:00*
-G04 #@! TF.ProjectId,RobotDriver,526f626f-7444-4726-9976-65722e6b6963,rev?*
-G04 #@! TF.SameCoordinates,Original*
-G04 #@! TF.FileFunction,Profile,NP*
-%FSLAX46Y46*%
-G04 Gerber Fmt 4.6, Leading zero omitted, Abs format (unit mm)*
-G04 Created by KiCad (PCBNEW (6.0.6-0)) date 2022-08-30 17:18:49*
-%MOMM*%
-%LPD*%
-G01*
-G04 APERTURE LIST*
-G04 #@! TA.AperFunction,Profile*
-%ADD10C,0.038400*%
-G04 #@! TD*
-G04 APERTURE END LIST*
-D10*
-X220000000Y-85000000D02*
-X195000000Y-122500000D01*
-X95000000Y-122500000D01*
-X95000000Y-99500000D01*
-X106500000Y-99500000D01*
-X106500000Y-69000000D01*
-X95000000Y-69000000D01*
-X95000000Y-47500000D01*
-X195000000Y-47500000D01*
-X220000000Y-85000000D01*
-M02*
