--- examples/test_robotdriver/gerbers/RobotDriver-PTH.drl
@@ -1,156 +1,0 @@
-M48
-; DRILL file {KiCad (6.0.6-0)} date 2022 August 30, Tuesday 17:18:55
-; FORMAT={-:-/ absolute / inch / decimal}
-; #@! TF.CreationDate,2022-08-30T17:18:55-07:00
-; #@! TF.GenerationSoftware,Kicad,Pcbnew,(6.0.6-0)
-; #@! TF.FileFunction,Plated,1,2,PTH
-FMAT,2
-INCH
-; #@! TA.AperFunction,Plated,PTH,ViaDrill
-T1C0.0118
-; #@! TA.AperFunction,Plated,PTH,ComponentDrill
-T2C0.0118
-; #@! TA.AperFunction,Plated,PTH,ComponentDrill
-T3C0.0394
-%
-G90
-G05
-T1
-X3.8976Y-4.1929
-X4.2421Y-2.3917
-X4.2815Y-2.4213
-X4.2913Y-2.8543
-X4.311Y-2.4606
-X4.311Y-3.7205
-X4.3406Y-3.937
-X4.3898Y-4.0748
-X4.4094Y-2.5984
-X4.4094Y-3.7795
-X4.4291Y-2.3327
-X4.5079Y-2.8937
-X4.5092Y-2.8629
-X4.5264Y-3.5815
-X4.5472Y-3.6024
-X4.5472Y-4.2323
-X4.6063Y-3.2677
-X4.626Y-4.1929
-X4.6654Y-3.7992
-X4.6654Y-4.1929
-X4.685Y-2.5394
-X4.7017Y-2.7923
-X4.7047Y-3.7205
-X4.7047Y-3.8386
-X4.7047Y-4.6457
-X4.7244Y-2.815
-X4.7244Y-3.5569
-X4.7256Y-3.7992
-X4.7441Y-3.0512
-X4.7511Y-3.529
-X4.765Y-3.0285
-X4.8228Y-3.2874
-X4.8425Y-3.4479
-X4.8622Y-3.4252
-X4.8622Y-3.9173
-X4.8819Y-3.9764
-X4.8831Y-3.8965
-X4.9594Y-3.7629
-X4.9803Y-3.4419
-X4.9803Y-3.7402
-X4.9803Y-4.1339
-X5.0Y-3.4646
-X5.0197Y-3.7205
-X5.0886Y-3.3563
-X5.1564Y-2.7264
-X5.1575Y-2.7559
-X5.1969Y-3.3465
-X5.2062Y-3.248
-X5.2152Y-4.145
-X5.2165Y-4.1142
-X5.2362Y-4.3307
-X5.2756Y-4.0945
-X5.315Y-2.4114
-X5.315Y-3.3661
-X5.3248Y-2.7756
-X5.3248Y-3.7205
-X5.3346Y-4.0945
-X5.371Y-4.1929
-X5.3937Y-2.3917
-X5.3937Y-4.2293
-X5.4331Y-3.7598
-X5.4497Y-3.7234
-X5.4539Y-3.7825
-X5.471Y-3.703
-X5.4724Y-2.7953
-X5.4921Y-3.6811
-X5.5118Y-2.5787
-X5.5118Y-2.7953
-X5.5118Y-3.9862
-X5.5512Y-4.3307
-X5.6693Y-4.252
-X5.7677Y-3.0512
-X5.7677Y-3.7598
-X5.7874Y-3.7371
-X5.7874Y-3.7825
-X5.7992Y-3.7096
-X5.8071Y-3.6811
-X5.8268Y-4.0748
-X5.8661Y-4.0521
-X5.8858Y-2.9134
-X5.8957Y-2.3228
-X6.0236Y-4.252
-X6.1614Y-3.6811
-X6.2008Y-4.311
-X6.3189Y-2.7953
-X6.3189Y-2.874
-X6.3189Y-3.0512
-X6.3583Y-3.5827
-X6.3583Y-3.9173
-X6.378Y-3.7598
-X6.5157Y-3.1496
-X6.5256Y-4.3406
-X6.6112Y-4.0748
-X6.6142Y-2.815
-X6.6142Y-2.9724
-X6.6339Y-3.0118
-X6.6535Y-2.815
-X6.6535Y-4.1142
-X6.6732Y-3.1496
-X6.7126Y-3.1496
-X6.752Y-3.8386
-X6.7913Y-3.7598
-X6.7999Y-2.767
-X6.8012Y-2.3917
-X6.8209Y-3.1496
-X6.8701Y-2.3917
-X6.9094Y-3.1988
-X6.9193Y-3.2283
-X6.939Y-3.0092
-X6.9488Y-2.7559
-X6.9685Y-3.0118
-X6.9697Y-2.735
-X6.9882Y-2.5492
-X6.998Y-3.4055
-X7.6772Y-3.7598
-X7.7264Y-4.0157
-X7.7362Y-2.4016
-X7.9429Y-2.815
-X8.2579Y-3.248
-T2
-X5.248Y-2.5512
-X5.248Y-2.6457
-X5.3425Y-2.5512
-X5.3425Y-2.6457
-X6.7244Y-2.5413
-X6.7244Y-2.6358
-X6.8189Y-2.5413
-X6.8189Y-2.6358
-T3
-X3.9862Y-2.2618
-X3.9862Y-2.3618
-X3.9862Y-2.4618
-X3.9862Y-2.5618
-X4.7638Y-2.1441
-X4.7638Y-2.2441
-X4.7638Y-2.3441
-T0
-M30
