--- examples/test_robotdriver/gerbers/RobotDriver-bottom-pos_jlc.csv
@@ -1,1 +1,0 @@
-Designator,Val,Package,Mid X,Mid Y,Rotation,Layer
