--- examples/test_robotdriver/gerbers/RobotDriver-top-pos_jlc.csv
@@ -1,65 +1,0 @@
-Designator,Val,Package,Mid X,Mid Y,Rotation,Layer
-1,,Indicator_LeadFree,186.000000,-91.500000,0.000000,top
-2,,Indicator_IdDots_4,186.250000,-95.000000,0.000000,top
-C1,C1779,C_0805_2012Metric,147.500000,-72.500000,-90.000000,top
-C2,C1779,C_0805_2012Metric,148.000000,-78.550000,90.000000,top
-C3,C15850,C_0805_2012Metric,110.250000,-98.000000,-90.000000,top
-C4,C14663,C_0603_1608Metric,108.250000,-98.000000,-90.000000,top
-C5,C15849,C_0603_1608Metric,112.250000,-98.000000,-90.000000,top
-C6,C14663,C_0603_1608Metric,139.000000,-102.000000,90.000000,top
-C7,C14663,C_0603_1608Metric,159.750000,-87.750000,0.000000,top
-C8,C14663,C_0603_1608Metric,159.750000,-85.750000,0.000000,top
-C9,C1779,C_0805_2012Metric,159.750000,-96.750000,0.000000,top
-C10,C14663,C_0603_1608Metric,159.750000,-90.250000,0.000000,top
-C11,C1779,C_0805_2012Metric,159.750000,-92.250000,0.000000,top
-C12,C14663,C_0603_1608Metric,159.750000,-98.750000,0.000000,top
-C13,C15850,C_0805_2012Metric,159.750000,-100.750000,0.000000,top
-C14,C14663,C_0603_1608Metric,176.500000,-87.250000,90.000000,top
-C15,C14663,C_0603_1608Metric,128.500000,-86.750000,180.000000,top
-C16,C14663,C_0603_1608Metric,128.500000,-83.750000,180.000000,top
-C17,C15850,C_0805_2012Metric,138.500000,-64.300000,-90.000000,top
-C18,C1779,C_0805_2012Metric,136.750000,-72.000000,180.000000,top
-C19,C57112,C_0603_1608Metric,133.000000,-72.250000,-90.000000,top
-C20,C15850,C_0805_2012Metric,176.000000,-63.750000,-90.000000,top
-C21,C1779,C_0805_2012Metric,174.250000,-71.750000,180.000000,top
-C22,C57112,C_0603_1608Metric,170.250000,-72.500000,-90.000000,top
-D1,C499796,D_SOD-123,122.500000,-83.500000,90.000000,top
-D2,C2920042,LED_WS2812B_PLCC4_5.0x5.0mm_P3.2mm,193.000000,-63.500000,90.000000,top
-D3,C2920042,LED_WS2812B_PLCC4_5.0x5.0mm_P3.2mm,198.500000,-74.000000,90.000000,top
-D4,C2920042,LED_WS2812B_PLCC4_5.0x5.0mm_P3.2mm,206.000000,-85.000000,90.000000,top
-D5,C2920042,LED_WS2812B_PLCC4_5.0x5.0mm_P3.2mm,198.500000,-93.000000,-90.000000,top
-D6,C2920042,LED_WS2812B_PLCC4_5.0x5.0mm_P3.2mm,192.500000,-104.500000,90.000000,top
-J1,C64658,JST_PH_B2B-PH-SM4-TB_1x02-1MP_P2.00mm_Vertical,153.250000,-57.250000,0.000000,top
-J2,,PinHeader_1x04_P2.54mm_Vertical,101.250000,-57.450000,0.000000,top
-J3,C265088,JST_PH_B6B-PH-SM4-TB_1x06-1MP_P2.00mm_Vertical,116.500000,-114.500000,180.000000,top
-J4,C265088,JST_PH_B6B-PH-SM4-TB_1x06-1MP_P2.00mm_Vertical,138.000000,-114.500000,180.000000,top
-J5,C265088,JST_PH_B6B-PH-SM4-TB_1x06-1MP_P2.00mm_Vertical,162.750000,-114.500000,180.000000,top
-J6,,Hirose_FH12-15S-0.5SH_1x15-1MP_P0.50mm_Horizontal,151.950000,-94.000000,-90.000000,top
-J7,C64658,JST_PH_B2B-PH-SM4-TB_1x02-1MP_P2.00mm_Vertical,140.500000,-57.250000,0.000000,top
-J8,C64658,JST_PH_B2B-PH-SM4-TB_1x02-1MP_P2.00mm_Vertical,129.000000,-57.250000,0.000000,top
-J9,C64658,JST_PH_B2B-PH-SM4-TB_1x02-1MP_P2.00mm_Vertical,178.750000,-57.250000,0.000000,top
-J10,C64658,JST_PH_B2B-PH-SM4-TB_1x02-1MP_P2.00mm_Vertical,166.750000,-57.250000,0.000000,top
-J11,,PinHeader_1x03_P2.54mm_Vertical,121.000000,-59.540000,180.000000,top
-L1,C98331,L_0805_2012Metric,151.500000,-78.500000,-90.000000,top
-R1,C23231,R_0603_1608Metric,158.000000,-73.000000,180.000000,top
-R2,C22809,R_0603_1608Metric,158.000000,-71.000000,0.000000,top
-R3,C25804,R_0603_1608Metric,117.250000,-98.000000,90.000000,top
-R4,C25804,R_0603_1608Metric,110.500000,-70.500000,-90.000000,top
-R5,C25804,R_0603_1608Metric,114.000000,-98.000000,90.000000,top
-R6,C23162,R_0603_1608Metric,137.500000,-85.000000,180.000000,top
-R7,C23162,R_0603_1608Metric,137.500000,-82.000000,180.000000,top
-R8,C23203,R_0603_1608Metric,159.750000,-94.750000,0.000000,top
-SW1,C318884,SW_SPST_SKQG_WithoutStem,101.000000,-106.250000,90.000000,top
-TP1,C238130,TestPoint_Keystone_5015_Micro-Minature,153.250000,-64.500000,180.000000,top
-TP2,C238130,TestPoint_Keystone_5015_Micro-Minature,112.500000,-57.000000,0.000000,top
-TP3,C238130,TestPoint_Keystone_5015_Micro-Minature,156.500000,-79.500000,0.000000,top
-TP4,C238130,TestPoint_Keystone_5015_Micro-Minature,142.000000,-83.000000,180.000000,top
-TP5,C238130,TestPoint_Keystone_5015_Micro-Minature,142.000000,-80.000000,180.000000,top
-U1,C500769,SOT-23-5,152.500000,-74.000000,270.0,top
-U2,C2934560,ESP-WROOM-02,113.500000,-84.000000,90.000000,top
-U3,C86832,SOIC-16W_7.5x10.3mm_P1.27mm,132.000000,-97.500000,180.0,top
-U4,C86832,SOIC-16W_7.5x10.3mm_P1.27mm,170.000000,-82.750000,180.0,top
-U5,C967633,Bosch_LGA-14_3x2.5mm_P0.5mm,132.000000,-85.000000,-90.000000,top
-U6,C50506,TSSOP-16-1EP_4.4x5mm_P0.65mm_EP3x3mm_ThermalVias,134.500000,-66.000000,180.0,top
-U7,C50506,TSSOP-16-1EP_4.4x5mm_P0.65mm_EP3x3mm_ThermalVias,172.000000,-65.750000,180.0,top
-U9,,Symbol_Duckling,185.968806,-83.691835,0.000000,top
