--- examples/test_robotdriver/gerbers/RobotDriver_jlc.csv
@@ -1,31 +1,0 @@
-Id,Designator,Package,Quantity,JLCPCB Part #,Supplier and ref,
-1,"C12,C14,C10,C4,C7,C6,C16,C15,C8",C_0603_1608Metric,9,C14663,,,
-2,"R6,R7",R_0603_1608Metric,2,C23162,,,
-3,R8,R_0603_1608Metric,1,C23203,,,
-4,"C19,C22",C_0603_1608Metric,2,C57112,,,
-5,"H1,H3,H2",JlcToolingHole_1.152mm,3,,,,
-6,"D4,D3,D5,D6,D2",LED_WS2812B_PLCC4_5.0x5.0mm_P3.2mm,5,C2920042,,,
-7,"C18,C2,C21,C1,C9,C11",C_0805_2012Metric,6,C1779,,,
-8,"J7,J10,J8,J9,J1",JST_PH_B2B-PH-SM4-TB_1x02-1MP_P2.00mm_Vertical,5,C64658,,,
-9,2,Indicator_IdDots_4,1,,,,
-10,J2,PinHeader_1x04_P2.54mm_Vertical,1,,,,
-11,1,Indicator_LeadFree,1,,,,
-12,U8,Symbol_LemurSmall,1,,,,
-13,C5,C_0603_1608Metric,1,C15849,,,
-14,U2,ESP-WROOM-02,1,C2934560,,,
-15,"TP1,TP2,TP4,TP5,TP3",TestPoint_Keystone_5015_Micro-Minature,5,C238130,,,
-16,J6,Hirose_FH12-15S-0.5SH_1x15-1MP_P0.50mm_Horizontal,1,,,,
-17,"J5,J4,J3",JST_PH_B6B-PH-SM4-TB_1x06-1MP_P2.00mm_Vertical,3,C265088,,,
-18,J11,PinHeader_1x03_P2.54mm_Vertical,1,,,,
-19,"R5,R3,R4",R_0603_1608Metric,3,C25804,,,
-20,"C20,C3,C17,C13",C_0805_2012Metric,4,C15850,,,
-21,"U3,U4",SOIC-16W_7.5x10.3mm_P1.27mm,2,C86832,,,
-22,U9,Symbol_Duckling,1,,,,
-23,SW1,SW_SPST_SKQG_WithoutStem,1,C318884,,,
-24,"U7,U6",TSSOP-16-1EP_4.4x5mm_P0.65mm_EP3x3mm_ThermalVias,2,C50506,,,
-25,R2,R_0603_1608Metric,1,C22809,,,
-26,D1,D_SOD-123,1,C499796,,,
-27,L1,L_0805_2012Metric,1,C98331,,,
-28,U1,SOT-23-5,1,C500769,,,
-29,R1,R_0603_1608Metric,1,C23231,,,
-30,U5,Bosch_LGA-14_3x2.5mm_P0.5mm,1,C967633,,,
