--- examples/test_robotdriver2/gerbers/RobotDriver2_Blimp-bottom-pos.csv
@@ -1,1 +1,0 @@
-Ref,Val,Package,PosX,PosY,Rot,Side
